refactor tool bar components for less coupling
